refactor(runtime): split node_vm.rs mini-interpreter under the 2000-line cap
Pure relocation, no functional changes: the deterministic vm.Script /
runIn*Context / compileFunction expression evaluator (split_top_level,
eval_expr, execute_statement, run_source and their helpers, lines
1014-1548) moves to new node_vm/eval.rs (545). node_vm.rs was 2258 after
this branch's +40-line death-pruning addition (2218 on main, already over
the cap); it is now 1727. run_source becomes pub(super); everything else
stays module-private via use super::*.
